terraform changes and provisioning cluster test with validation
This change includes helm installation of all the components
needed to test the payg billing adapter. It includes
csp-adapter-crd, csp-rancher-usage-operator and the
rancher-csp-billing-adapter charts.

Also includes a bash script to generate the rancher
admin token and creates a cattle-config-e2e.yaml file
with all the contents specific to the deployment and
the configuration entries that the tests need.

Also includes a new set of tests which provisions a
hosted eks cluster and validates that the managed_node_count
that the custom resource csprancherusagerecords reports is
correct.

Note: This commit does not install the old byol csp adapter.